requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
crewai>=0.11.0
fastapi-mcp>=0.3.0
python-dotenv==1.0.0
langchain>=0.0.350
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 